bug tool: Add initial version
Create a simplified 'bug' utility with just three commands:

  bug version
  bug help
  bug test

Where 'bug test' checks a silly thing: whether the 'getpid()'
system call works fine.

This utility can now be extended to test for specific kernel bugs,
and to make sure we don't regress the same bugs in the future.

The code is derived from tools/perf/ which is in turn derived
from the Git project.

Signed-off-by: Ingo Molnar <mingo@elte.hu>